Moved all verilog-specific syntax to WriteVLOGSyntax.py, and made generator_verilog.py (renamed generator_hdl.py) and WriteHDLUtils.py compatible with verilog and VHDL (once WriteVHDLSyntax.ph gets written)
