+++ generator_hdl.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python
 
 from TrackletGraph import MemModule, ProcModule, TrackletGraph
-from WriteHDLUtils import getHLSMemoryClassName, groupAllConnectedMemories, writeTBMemoryStimulusInstance, writeTBMemoryReadInstance, writeTopLevelMemoryInstance, writeControlSignals_interface, writeMemoryLHSPorts_interface, writeMemoryRHSPorts_interface, writeTBControlSignals, writeFWBlockControlSignals, writeFWBlockMemoryLHSPorts, writeFWBlockMemoryRHSPorts, writeModuleInstance
+from WriteHDLUtils import groupAllConnectedMemories, writeModuleInstance
+from WriteVLOGSyntax import writeTopModuleOpener, writeTBOpener, writeModuleCloser, writeTimescale, writeTBMemoryStimulusInstance, writeTBMemoryReadInstance, writeTopLevelMemoryInstance, writeControlSignals_interface, writeMemoryLHSPorts_interface, writeMemoryRHSPorts_interface, writeTBControlSignals, writeFWBlockControlSignalPorts, writeFWBlockMemoryLHSPorts, writeFWBlockMemoryRHSPorts
 import os, subprocess
 
 ########################################
@@ -25,7 +26,7 @@ def writeMemoryModules(mem_list, interface=0):
     string_mem = ""
     # Loop over memories in the list
     for memModule in sorted(mem_list,key=lambda x: x.index):
-        amem_str,_ = writeTopLevelMemoryInstance(memModule,interface)
+        amem_str = writeTopLevelMemoryInstance(memModule,interface)
         string_mem += amem_str
         
     return string_mem
@@ -82,7 +83,7 @@ def writeTopModule_interface(topmodule_name, process_list, memories_topin, memor
         if proc.is_first: initial_proc = proc.mtype
         if proc.is_last: final_proc = proc.mtype
 
-    string_topmod_interface = "module "+topmodule_name+"(\n"
+    string_topmod_interface = writeTopModuleOpener(topmodule_name)
 
     # Write control signals
     string_ctrl_signals = writeControlSignals_interface(initial_proc, final_proc)
@@ -143,11 +144,11 @@ def writeTopFile(topfunc, process_list, memList_topin, memList_inside, memlist_t
                               memList_topin, memList_topout)
 
     string_src = ""
-    string_src += "`timescale 1ns / 1ps\n\n"
+    string_src += writeTimescale()
     string_src += string_topmod_interface
     string_src += string_memModules
     string_src += string_procModules
-    string_src += "\n\nendmodule"
+    string_src += writeModuleCloser()
 
     return string_src
 
@@ -181,7 +182,7 @@ def writeTBMemoryReads(memories_list, emData_dir="", sector="04"):
 def writeFWBlockInstance(topfunc, memories_in, memories_out, first_proc, last_proc):
     string_inmems_ports = ""
 
-    string_fwblock_ctrl = writeFWBlockControlSignals(first_proc, last_proc)
+    string_fwblock_ctrl = writeFWBlockControlSignalPorts(first_proc, last_proc)
 
     for memModule in memories_in:
         string_inmems_ports += writeFWBlockMemoryLHSPorts(memModule)
@@ -218,9 +219,9 @@ def writeTestBench(topfunc, memories_in, memories_out, emData_dir, sector="04"):
             break
 
     # Write test bench header
-    string_header = "// Test bench generated by generator_verilog.py\n\n"
-    string_header += "`timescale 1ns / 1ps\n\n"
-    string_header += "module " + topfunc + "_test();\n\n"
+    string_header = ""
+    string_header += writeTimescale()
+    string_header += writeTBOpener(topfunc)
 
     string_ctrl_signals = writeTBControlSignals(topfunc, first_proc, last_proc)
 
@@ -235,7 +236,7 @@ def writeTestBench(topfunc, memories_in, memories_out, emData_dir, sector="04"):
     string_tb += string_memin
     string_tb += string_memout
     string_tb += string_fwblock_inst
-    string_tb += "\n\nendmodule"
+    string_tb += writeModuleCloser()
     
     return string_tb,""
